Add functions to separate file
